test_requirements.txt
coverage
django-coverage
fabric3
flake8
ipdb
mixer
mock
tox
